adding support for WikiTeam dumps repository;
git-svn-id: https://forja.rediris.es/svn/statmediawiki@318 25f239b6-92f9-4231-b61e-0ce2aa185c45

+++ branches/np/smwparser.py
@@ -16,7 +16,9 @@
 # along with this program.  If not, see <http://www.gnu.org/licenses/>.
 
 import datetime
+import re
 import sqlite3
+import subprocess
 import time
 import xmlreader
 import sys
@@ -83,6 +85,13 @@ def generateAuxTables(conn=None, cursor=None):
     generateUserTable(conn=conn, cursor=cursor)
 
 def parseMediaWikiXMLDump(dumpfilename, dbfilename):
+    #fix: verificar que dumpfilename solo tiene un fichero, el .xml, en otro caso, abortar
+    s = subprocess.Popen('7z l %s' % dumpfilename, shell=True, stdout=subprocess.PIPE, bufsize=65535).stdout
+    if not re.search(ur'(?m)1 files, 0 folders\n$', s.read()):
+        print 'ERROR: the file %s contains more files than a .xml' % dumpfilename
+        return 
+        #os.rename(filename, '%s.corrupted' % filename)
+    
     if os.path.exists(dbfilename): #si existe lo borramos, pues el usuario ha marcado sobreescribir, sino no entraría aquí
         os.remove(dbfilename)
